Ajustes visuales y nuevo elemento starRating

+++ AvengersFundaments/Feature/Domain/Models/Character.swift
@@ -12,10 +12,10 @@ class Character {
     var name: String
     var lastname: String
     var image: String?
-    var score: Float = 0.0
+    var score: Double = 0.0
     var description: String
     
-    init(name: String, lastname: String, image: String, score: Float, description: String) {
+    init(name: String, lastname: String, image: String, score: Double, description: String) {
         self.name = name
         self.lastname = lastname
         self.image = image
